can set max number of views to show on carousel
If you have too many views (say over 1000), having them all loaded into the carousel takes a long time and make the scrolling choppy.  Now you can specify a maximum number of items to show.  As the user scrolls, the carousel will ask the datasource for views when they are needed.  Views are kept in a cache so that the datasource does not have to recreate it every time.