fixed scrolling issue: click while carousel is decelerating causes it to jump backwards
After dragging the carousel from right to left quickly and releasing the mouse, the carousel begins to decelerate.  If the mouse is pressed down during deceleration, the carousel will instantly jump back to the first item deceleration began on.  This is due to startVelocity being set to 0 on mouseDown, and thus the deceleration logic calculates there has been no movement at all returning to the original position.

The desired effect would be to have the carousel stop where it is.  The user could then drag, or release allowing the carousel to scroll to the nearest item.  This would be a smoother experience.

I set scrolling and deceleration to NO in mouseDown instead of mouseDragged, so that the deceleration would be avoided immediately after a mouseDown, preventing the jump backwards.