if mouseUp has clickCount 1, only process click, do nothing else
this prevents the carousel from initiating deceleration/scroll when the user just clicks it, which would result in an deceleration/scrollDidEnd msg being sent to the delegate, which is usually not what we want from just a click.  I think it is possible for the user to drag slightly and have it still count as a click, which means the carousel would be slightly off from where it's supposed to "snap" to, but not a big deal for now.