reverting previous change.. mouseUp performs scroll action even if the event was registered as a click
Otherwise, if the user clicks while scrolling the carousel quickly, it will stop midway, looking awkward.  Also the didEndScrolling msg is passed to delegates even when the carousel view is resized.  So the best way right now for the delegate to handle the msg appropriately is to keep track of the last index of the carousel, and determine whether there's been a change to the index.